==> notebooks/02_data_collection.ipynb <==
## Listing needed Data

The data set includes information about:

- **Customers who left within the last month**: the column is called Churn
- **Signed up Services**: phone, multiple lines, internet, online security, online backup, device protection, tech support, and streaming TV and movies.
- **Account information**: how long they’ve been a customer, contract, payment method, paperless billing, monthly charges, and total charges.
- **Demographic info**: gender, age range, and if they have partners and dependents.

## Data Size Estimate:

Ideally, several thousand records (customers) are needed for reliable modeling.<br>
**Minimum**: At least 1,000 customers' worth of data, depending on the complexity and features.

## Data Availability

Data is available at <a href="https://www.kaggle.com/datasets/blastchar/telco-customer-churn">kaggle</a> 

## Space it takes

Too smalll ~1mb.

## Legal obligations & Authorization

Its open-source data for any use.

## Get Data

Download data from kaggle should be located at **../data/WA_Fn-UseC_-Telco-Customer-Churn.csv**

## Load Data

import numpy as np
import pandas as pd

# Load CSV Data
data = pd.read_csv("../data/WA_Fn-UseC_-Telco-Customer-Churn.csv")
data.head()

# Shape of the data
data.shape

# Check Balance of the data
np.unique(data.Churn,return_counts=True)

We see that data is not balanced at all.<br>
**SMOTE** will be used to balance it later.

## Check the size and type of data

It’s a sample of records of customer interactions and account details.<br>
Contains 7_000 customer records.

## Sample a test set

from sklearn.model_selection import train_test_split
train_data,test_data = train_test_split(data,stratify=data.Churn,test_size=0.1,random_state=42)

train_data.to_csv("../data/train_data.csv",index=False)
test_data.to_csv("../data/test_data.csv",index=False)
