==> notebooks/04_data_preparation.ipynb <==
import numpy as np
import pandas as pd

from prefect import task,flow
from joblib import dump,load

## Load Processed Train Data

data = pd.read_csv("../data/train_processed_data.csv")
data.drop(columns=["Unnamed: 0"],inplace=True)
data.info()

X,y = data.drop("Churn",axis=1),data["Churn"]

# Build Load Task
@task
def load_data(data):
    data = pd.DataFrame(data,columns=['gender', 'SeniorCitizen', 'Partner', 'Dependents', 'tenure',
       'PhoneService', 'MultipleLines', 'InternetService', 'OnlineSecurity',
       'OnlineBackup', 'DeviceProtection', 'TechSupport', 'StreamingTV',
       'StreamingMovies', 'Contract', 'PaperlessBilling', 'PaymentMethod',
       'MonthlyCharges', 'TotalCharges'])
    return data

## Feature Selection

**CustomerID** is an identity feature with no benefit for the task.<br>
So it will get dropped.

data.drop(columns=["customerID"],inplace=True)

## Feature Cleaning

data.isnull().sum()

We see that **Total Charges** has
7 missing values.

from sklearn.impute import SimpleImputer

# Fill with Mode
mode_features = ["gender","SeniorCitizen","Partner","Dependents","PhoneService","MultipleLines","InternetService","OnlineSecurity","OnlineBackup","DeviceProtection","TechSupport","StreamingTV","StreamingMovies","Contract","PaperlessBilling","PaymentMethod"]
mode_imputer = SimpleImputer(strategy="most_frequent")
mode_imputer.fit(data[mode_features])

dump(mode_imputer,"../pipelines/mode_imputer.joblib")

# Fill with Mean
mean_features = ["tenure","MonthlyCharges","TotalCharges"]
mean_imputer = SimpleImputer(strategy="mean")
mean_imputer.fit(data[mean_features])

dump(mean_imputer,"../pipelines/mean_imputer.joblib")

# Build The Clean Task
@task
def clean_data(org_data):
    data = org_data.copy()
    mode_features = ["gender","SeniorCitizen","Partner","Dependents","PhoneService","MultipleLines","InternetService","OnlineSecurity","OnlineBackup","DeviceProtection","TechSupport","StreamingTV","StreamingMovies","Contract","PaperlessBilling","PaymentMethod"]
    mean_features = ["tenure","MonthlyCharges","TotalCharges"]

    mode_cleaned_data = pd.DataFrame(mode_imputer.transform(data[mode_features]),columns=mode_features)
    mean_cleaned_data = pd.DataFrame(mean_imputer.transform(data[mean_features]),columns=mean_features)

    return pd.concat([mode_cleaned_data,mean_cleaned_data],axis=1)

## Feature Engineering

Create **new features** from the **existing ones**.

# Build The Feature Engineering Task
@task
def feature_engineering(org_data):
    data = pd.DataFrame(org_data,columns=['gender', 'SeniorCitizen', 'Partner', 'Dependents', 'tenure',
   'PhoneService', 'MultipleLines', 'InternetService', 'OnlineSecurity',
   'OnlineBackup', 'DeviceProtection', 'TechSupport', 'StreamingTV',
   'StreamingMovies', 'Contract', 'PaperlessBilling', 'PaymentMethod',
   'MonthlyCharges', 'TotalCharges'])
        
    data['TenureCategory'] = pd.cut(data['tenure'], bins=[-1,3,12,np.inf],labels=["Recent", "Established", "Loyal"]) 
    data['Tenure_X_MonthlyCharges'] = data['tenure'] * data['MonthlyCharges']
    data['HighSpender'] = data['TotalCharges'].apply(lambda x: True if x > 5000 else False)
    data['MonthlyChargesRatio'] = np.where(data['TotalCharges']!=0,data['MonthlyCharges'] / data['TotalCharges'],0)
    data['SqrtTotalCharges'] = data['TotalCharges'].apply(np.sqrt)
    data['SqrtMonthlyCharges'] = data['MonthlyCharges'].apply(np.sqrt)
    data['AutomaticPayment'] = data['PaymentMethod'].apply(lambda x: True if "automatic" in x else False)
    data['AllServicesActivated'] =  (data['OnlineSecurity'] == "Yes") & (data['OnlineBackup'] == "Yes") & \
                                   (data['DeviceProtection'] == "Yes") & (data['TechSupport'] == "Yes")
    data['TeleServicesActivated'] = (data['PhoneService'] == "Yes") & (data['MultipleLines'] == "Yes") & \
                                    (data['InternetService'] == "Yes")
    data['Connected'] = (data['Partner'] == "Yes") & (data['Dependents'] == "Yes")
    
    return data

processed_data = feature_engineering(clean_data(load_data(X)))

## Numerical Feature Scaling

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
scaler.fit(processed_data[["tenure","MonthlyCharges","TotalCharges","MonthlyChargesRatio","Tenure_X_MonthlyCharges","SqrtTotalCharges","SqrtMonthlyCharges"]])

dump(scaler,"../pipelines/scaler.joblib")

# Build The Scaling Task
num_features = ["tenure","MonthlyCharges","TotalCharges","MonthlyChargesRatio","Tenure_X_MonthlyCharges","SqrtTotalCharges","SqrtMonthlyCharges"]

@task
def scaling_data(data):
    data[num_features] = scaler.transform(data[num_features])

    return data

## Categorical Feature Encoding

from sklearn.preprocessing import OrdinalEncoder,OneHotEncoder

ordinal_features = ["SeniorCitizen","Partner","Dependents","PhoneService","MultipleLines","InternetService","OnlineSecurity","OnlineBackup","DeviceProtection","TechSupport","StreamingTV","StreamingMovies","Contract","PaperlessBilling","TenureCategory","HighSpender","AutomaticPayment","AllServicesActivated","TeleServicesActivated","Connected"]

binary = [False,True]
binary_yes_no = ["No","Yes"]
multi = ["No internet service","No","Yes"]
custom_categories = [
    binary_yes_no,
    binary_yes_no,
    binary_yes_no,
    binary_yes_no,
    binary_yes_no,
    ["No","DSL","Fiber optic"],
    multi,
    multi,
    multi,
    multi,
    multi,
    multi,
    ["Month-to-month","One year","Two year"],
    binary_yes_no,
    ["Recent", "Established", "Loyal"],
    binary,
    binary,
    binary,
    binary,
    binary,
]
ord_enc = OrdinalEncoder(categories=custom_categories)
ord_enc.fit(processed_data[ordinal_features])

dump(ord_enc,"../pipelines/ord_enc.joblib")

nominal_features = ["gender","PaymentMethod"]
one_hot_enc = OneHotEncoder(sparse_output=False)
one_hot_enc.fit(processed_data[nominal_features])

dump(one_hot_enc,"../pipelines/one_hot_enc.joblib")

# Build The Encoding Task
@task
def encoding_data(org_data):
    data = org_data.copy()
    ordinal_features = ["SeniorCitizen","Partner","Dependents","PhoneService","MultipleLines","InternetService","OnlineSecurity","OnlineBackup","DeviceProtection","TechSupport","StreamingTV","StreamingMovies","Contract","PaperlessBilling","TenureCategory","HighSpender","AutomaticPayment","AllServicesActivated","TeleServicesActivated","Connected"]
    nominal_features = ["gender","PaymentMethod"]

    data[ordinal_features] = ord_enc.transform(data[ordinal_features])
    
    one_hot_encoded_data = one_hot_enc.transform(data[nominal_features])
    one_hot_encoded_data_columns = one_hot_enc.get_feature_names_out(input_features=data[nominal_features].columns)
    one_hot_data = pd.DataFrame(one_hot_encoded_data,columns=one_hot_encoded_data_columns)
    
    return pd.concat([data,one_hot_data],axis=1).drop(nominal_features,axis=1)

## Build the Data Pipeline

@flow(name="Pipeline")
def pipeline(data):
    processed_data = load_data(data)
    processed_data = clean_data(processed_data)
    processed_data = feature_engineering(processed_data)
    processed_data = scaling_data(processed_data)
    processed_data = encoding_data(processed_data)
    return processed_data

# Encode Target Feature
from sklearn.preprocessing import LabelEncoder

y = pd.Series(LabelEncoder().fit_transform(y),name="Churn")

processed_data = pd.concat([pipeline(X),y],axis=1)
processed_data

processed_data.to_csv("../data/train_processed_data.csv")
