==> notebooks/06_finetune_models.ipynb <==
import numpy as np
import pandas as pd

from pycaret.classification import *

from joblib import load,dump

# Load Training Processed Data

data = pd.read_csv("../data/train_processed_data.csv")
data.drop("Unnamed: 0",axis=1,inplace=True)
data

X,y = data.drop("Churn",axis=1),data.Churn

# Setup PyCaret

# Setup PyCaret AutoML
clf = setup(data=data,target="Churn",preprocess=False,train_size=0.9)

# Get Validation Data from PyCaret Setup
X_valid,y_valid = get_config("X_test"),get_config("y_test")

# Fine Tunning Base Models 

# Load the models
models = ["AdaBoostClassifier","LinearDiscriminantAnalysis","GaussianNB","LogisticRegression","GradientBoostingClassifier"]
for i in range(len(models)):
    models[i] = load(f"../models/{models[i]}.joblib")

# Fine Tunning
best_models = []
for model in models:
    best_models.append(tune_model(model,fold=3))

# Store Fine Tuned Models
for model in best_models:
    dump(model,f"../models/finetuned/best_{model.__class__.__name__}")

# Ensemble Methods

## Voting

from sklearn.ensemble import VotingClassifier

voting_clf = VotingClassifier(estimators=list(zip(["AdaBoost","LDA","GaussNB","Logistic","GBoost"],best_models)),voting="soft")
voting_clf.fit(X,y)

# Performace of the Voting Classifier
from sklearn.metrics import accuracy_score,precision_score,recall_score,f1_score

y_pred = voting_clf.predict(X_valid)
print("Accuracy Score: ",accuracy_score(y_pred,y_valid))
print("Precision Score: ",precision_score(y_pred,y_valid))
print("Recall Score: ",recall_score(y_pred,y_valid))
print("F1 Score: ",f1_score(y_pred,y_valid))

# Store it
dump(voting_clf,"../models/finetuned/voting_clf.joblib")

## Stacking

from sklearn.ensemble import StackingClassifier

stacking_clf = StackingClassifier(estimators=list(zip(["AdaBoost","LDA","GaussNB","Logistic"],best_models[:-1])),final_estimator=best_models[-1])
stacking_clf.fit(X,y)

# Performace of the Stacking Classifier
y_pred = stacking_clf.predict(X_valid)
print("Accuracy Score: ",accuracy_score(y_pred,y_valid))
print("Precision Score: ",precision_score(y_pred,y_valid))
print("Recall Score: ",recall_score(y_pred,y_valid))
print("F1 Score: ",f1_score(y_pred,y_valid))

# Store it
dump(stacking_clf,"../models/finetuned/stacking_clf.joblib")

# Best Model

**Voting Classifier** has a higher F1 score and faster,so its the chosen one.

# Check its performance on the Test Set
test_data = pd.read_csv("../data/test_data.csv")
test_data

Process it using the **Pipeline**<br>
We will use python file for this cause pipeline is saved in python file and notebook file cant interacte with it.<br>

Check **scripts/process_test_data.py**

test_processed_data = pd.read_csv("../data/test_processed_data.csv")
# Add Churn to the processed test data from the original test data
from sklearn.preprocessing import LabelEncoder
test_processed_data["Churn"] = LabelEncoder().fit_transform(pd.read_csv("../data/test_data.csv")["Churn"])

X_test,y_test = test_processed_data.drop("Churn",axis=1),test_processed_data["Churn"]

# Performace On the Test Data
y_pred = voting_clf.predict(X_test)
print("Accuracy Score: ",accuracy_score(y_pred,y_test))
print("Precision Score: ",precision_score(y_pred,y_test))
print("Recall Score: ",recall_score(y_pred,y_test))
print("F1 Score: ",f1_score(y_pred,y_test))

**Acceptable**
